Crouching is now smooth through projectsettings change, fixed bug where mashing shift would cause the player to get stuck in the air
Ticked 'Auto Sync Transforms' setting in the Project Settings for Physics, which fixed the janky get-up for crouching.

Also removed the code where crouching moved the ground-check up, as the player's bottom now sticks to the ground, so it allowed the player to get stuck in the air, unable to jump or crouch when shift is spammed.