Tap jump should work on all systems, can no longer float and jump in charge shot
Made tap jump script make use of Time.deltaTime  so it should be consistent between all systems.

Added a tag called 'NotGround', which we can tag with anything we don't want the player to be able to jump on/in.

I've added this to the charge shot effect so the player can no longer jump while in its radius.