BIG MOVEMENT PATCH, Fixed shorthop, dashes now cancel, player no longer floats while mashing dash upward momentum,
Shorthop was fixed - the minimum jump value was playing multiple times over which often made the player reach the peak height on even a small jump press.

Dashes now set player's y velocity to 0 and set jump height to 0

made sure that enable/desable dash function (which contains the set gravity to 0 part of the dash code) doesn't work if the player has used all their dashes

Jump height is still inconsistent but not as much as before, tried a lot of different things and will come back to it soon.