Fixed infinite speed bug, running faster than sliding bug and updated script formatting
SerializeFields are now on the same line to make the script more readable.

The velocity applied to the player now has 'move' subtracted from it, so they don't travel faster moving after a slide with slide momentum than sliding alone.

While easily fixable, there is a BRIEF window where if the player jumps they move faster. I really want to keep this, or something like this in!