Fixed sliding under world bug, optimised head & ground check
I forgot to reset a bool that auto-crouches the player when finishing a slide under a small gap, meaning the auto-crouch was only done once. Reset it after a slide is finished so this save happens consistently.

Additionally made head & ground check dynamic to player height, allowing me to remove the code that shifts their position upon crouching or sliding.