Many playerMove bugs fixed
- Slight drifting bug has been entirely fixed
- bug where player returned to regular slope interactions has been fixed
- Huge optimisation to momentum formula
- Modular distance added to slopecheck in editor

TO NOTE:
Due to the momentum formula change required to fix the drifting bug, the momentum's power was cut hugely. To fix this, I've significantly upped the slidestrength and momentum strength values. There may be some changes to the player's feel, but I've tried my damndest to get it as close as possible to how it was before through rigorous level testing.